requirements.txt
rembg[cpu]==2.0.75
Pillow>=10.0.0
requests>=2.31.0
